ai-service/quiz-generator/requirements.txt
openai>=1.12.0
pydantic>=2.6.0
python-dotenv>=1.0.0
streamlit>=1.31.0
tenacity>=8.2.0
